+++ Code/GUI/lib/QGoTabImageView4D.cxx
@@ -1097,23 +1097,31 @@ QGoTabImageView4D::ChangeContourRepresentationProperty()
   QColor activenodecolor; // =
                           // m_ManualSegmentationWidget->GetActiveNodesColor();
 
-  double rl, gl, bl;
-
+  qreal rl, gl, bl;
   linecolor.getRgbF(&rl, &gl, &bl);
 
-  double rn, gn, bn;
+  qreal rn, gn, bn;
   nodecolor.getRgbF(&rn, &gn, &bn);
 
-  double ra, ga, ba;
+  qreal ra, ga, ba;
   activenodecolor.getRgbF(&ra, &ga, &ba);
 
   for ( unsigned int i = 0; i < m_ContourRepresentation.size(); i++ )
     {
     m_ContourRepresentation[i]->GetLinesProperty()->SetLineWidth(linewidth);
-    m_ContourRepresentation[i]->GetLinesProperty()->SetColor(rl, gl, bl);
+    m_ContourRepresentation[i]->GetLinesProperty()->SetColor(
+      static_cast< double >( rl ),
+      static_cast< double >( gl ),
+      static_cast< double >( bl ) );
 
-    m_ContourRepresentation[i]->GetProperty()->SetColor(rn, gn, bn);
-    m_ContourRepresentation[i]->GetActiveProperty()->SetColor(ra, ga, ba);
+    m_ContourRepresentation[i]->GetProperty()->SetColor(
+      static_cast< double >( rn ),
+      static_cast< double >( gn ),
+      static_cast< double >( bn ) );
+    m_ContourRepresentation[i]->GetActiveProperty()->SetColor(
+      static_cast< double >( ra ),
+      static_cast< double >( ga ),
+      static_cast< double >( ba ) );
     }
 }
